Refact tests to extract duplicated code
