+++ MyMoviesTests/Model/TrendingViewModelSpec.swift
@@ -9,36 +9,12 @@ import XCTest
 @testable import MyMovies
 
 final class TrendingViewModelSpec: XCTestCase {
-
-    private func getMockData(for path: String) -> Data? {
-        guard let pathString = Bundle(for: type(of: self)).path(forResource: path, ofType: "json") else {
-            fatalError("Mock json not found")
-        }
-
-        guard let jsonString = try? String(contentsOfFile: pathString, encoding: .utf8) else {
-            fatalError("Mock json can not be converted to String")
-        }
-
-        return jsonString.data(using: .utf8)
-    }
-
-    private func getMockJson(for data: Data) -> TrendingResult {
-        let jsonDecoder = JSONDecoder()
-        jsonDecoder.keyDecodingStrategy = .convertFromSnakeCase
-        do {
-            let body = try jsonDecoder.decode(TrendingResult.self, from: data)
-            return body
-        } catch {
-            fatalError("Error when decoding json")
-        }
-    }
-
     func testSuccessJson() {
-        guard let data = getMockData(for: "trending-result") else {
+        guard let data = JsonHelper().getMockData(for: "trending-result") else {
             fatalError("Mock json can not be converted to Data")
         }
 
-        let trending = getMockJson(for: data)
+        let trending = JsonHelper().getMockJson(for: data)
         let viewModel = TrendingViewModel(trending: trending)
 
         XCTAssertEqual(viewModel.title, trending.originalTitle)
@@ -47,11 +23,11 @@ final class TrendingViewModelSpec: XCTestCase {
     }
 
     func testNullableJson() {
-        guard let data = getMockData(for: "trending-result-nullable") else {
+        guard let data = JsonHelper().getMockData(for: "trending-result-nullable") else {
             fatalError("Mock json can not be converted to Data")
         }
 
-        let trending = getMockJson(for: data)
+        let trending = JsonHelper().getMockJson(for: data)
         let viewModel = TrendingViewModel(trending: trending)
 
         XCTAssertEqual(viewModel.title, "")
